HSB(candisc)	R Documentation

High School and Beyond Data

Description

The High School and Beyond Project was a longitudinal study of students in the U.S. carried out in 1980 by the National Center for Education Statistics. Data were collected from 58,270 high school students (28,240 seniors and 30,030 sophomores) and 1,015 secondary schools. The HSB data frame is sample of 600 observations, of unknown characteristics, originally taken from Tatsuoka (1988).

Usage

data(HSB)

Format

A data frame with 600 observations on the following 15 variables. There is no missing data.

id

Observation id: a numeric vector

gender

a factor with levels male female

race

Race or ethnicity: a factor with levels hispanic asian african-amer white

ses

Socioeconomic status: a factor with levels low middle high

sch

School type: a factor with levels public private

prog

High school program: a factor with levels general academic vocation

locus

Locus of control: a numeric vector

concept

Self-concept: a numeric vector

mot

Motivation: a numeric vector

career

Career plan: a factor with levels clerical craftsman farmer homemaker laborer manager military operative prof1 prof2 proprietor protective sales school service technical not working

read

Standardized reading score: a numeric vector

write

Standardized writing score: a numeric vector

math

Standardized math score: a numeric vector

sci

Standardized science score: a numeric vector

ss

Standardized social science (civics) score: a numeric vector

Source

Tatsuoka, M. M. (1988). Multivariate Analysis: Techniques for Educational and Psychological Research (2nd ed.). New York: Macmillan, Appendix F, 430-442.

Retrieved from: http://www.gseis.ucla.edu/courses/data/hbs6.dta

References

High School and Beyond data files: http://www.sscnet.ucla.edu/issr/da/index/techinfo/I78961.HTM

Examples


str(HSB)
# main effects model
hsb.mod <- lm( cbind(read, write, math, sci, ss) ~
                gender + race + ses + sch + prog, data=HSB)
Anova(hsb.mod)

# Add some interactions
hsb.mod1 <- update(hsb.mod, . ~ . + gender:race + ses:prog)
heplot(hsb.mod1, col=palette()[c(2,1,3:6)], variables=c("read","math"))

hsb.can1 <- candisc(hsb.mod1, term="race")
heplot(hsb.can1, col=c("red", "black"))

# show canonical results for all terms
hsb.can <- candiscList(hsb.mod)
hsb.can


[Package candisc version 0.5-10 Index]