candiscList(candisc)	R Documentation

Canonical discriminant analyses

Description

candiscList performs a generalized canonical discriminant analysis for all terms in a multivariate linear model (i.e., an mlm object), computing canonical scores and vectors.

Usage


candiscList(mod, ...)

## S3 method for class 'mlm':
candiscList(mod, type = "2", manova, ndim, ...)

## S3 method for class 'candiscList':
plot(x, term, ask = interactive(), graphics = TRUE, ...)

## S3 method for class 'candiscList':
print(x, ...)

## S3 method for class 'candiscList':
summary(object, ...)

Arguments

mod 	 An mlm object, such as computed by lm() with a multivariate response

type 	 type of test for the model term, one of: "II", "III", "2", or "3"

manova 	 the Anova.mlm object corresponding to mod. Normally, this is computed internally by Anova(mod)

ndim 	 Number of dimensions to store in the means, structure, scores and coeffs.* components. The default is the rank of the H matrix for the hypothesis term.

object, x 	 A candiscList object

term 	 The name of one term to be plotted for the plot method. If not specified, one candisc plot is produced for each term in the mlm object.

ask 	 If TRUE (the default, when running interactively), a menu of terms is presented; if ask is FALSE, canonical plots for all terms are produced.

graphics

... 	 arguments to be passed down.

Value

An object of class candiscList which is a list of candisc objects for the terms in the mlm.

Author(s)

Michael Friendly and John Fox

See Also

candisc, heplot, heplot3d

Examples


grass.mod <- lm(cbind(N1,N9,N27,N81,N243) ~ Block + Species, data=Grass)

grass.canL <-candiscList(grass.mod)
print(grass.canL)
plot(grass.canL, type="n", ask=FALSE)


[Package candisc version 0.5-10 Index]