Breakfast Cereal Dataset

A multivariate dataset describing seventy-seven commonly available breakfast cereals, based on the information now available on the FDA food label. The variable rating is a likely response variable in statistical models.

Usage

data("cereal")

Format

A data frame with 77 observations on the following 16 variables.

name

cereal name, a character vector

mfr

manufacturer (A G K N P Q R), a character vector

type

type (cold/hot), a character vector

calories

calories (number), a numeric vector

protein

protein(g), a numeric vector

fat

fat(g), a numeric vector

sodium

sodium(mg), a numeric vector

fiber

dietary fiber(g), a numeric vector

carbo

complex carbohydrates(g), a numeric vector

sugars

sugars(g), a numeric vector

potass

potassium(mg), a numeric vector

vitamins

vitamins & minerals (0, 25, or 100, respectively indicating "none added"; "enriched, often to 25% FDA recommended"; "100% of FDA recommended"), a numeric vector

shelf

display shelf (1, 2, or 3, counting from the floor), a numeric vector

weight

weight (in ounces) of one serving (serving size), a numeric vector

cups

cups per serving, a numeric vector

rating

health rating of the cereal (unknown calculation method), a numeric vector

Source

From the American Statistical Association 1993 Statistical Graphics Exposition, 'Serial Correlation or Cereal Correlation ??', https://community.amstat.org/jointscsg-section/dataexpo/dataexpo1993.

Details

This dataset was used in the poster competition for the American Statistical association 1993 Statistical Graphics Exposition, titled Serial Correlation or Cereal Correlation ??.

The call for participation reads: "A multivariate dataset describing seventy-seven commonly available breakfast cereals, based on the information now available on the newly-mandated F&DA food label. What are you getting when you eat a bowl of cereal? Can you get a lot of fiber without a lot of calories? Can you describe what cereals are displayed on high, low, and middle shelves? The good news is that none of the cereals for which we collected data had any cholesterol, and manufacturers rarely use artificial sweeteners and colors, nowadays. However, there is still a lot of data for the consumer to understand while choosing a good breakfast cereal."

Further details on the variables and suggested analyses are available at https://community.amstat.org/jointscsg-section/dataexpo/dataexpo1993

See also

Examples

data(cereal)
## None yet
## maybe str(cereal) ; plot(cereal) ...