Grass(candisc)	R Documentation

Yields from Nitrogen nutrition of grass species

Description

The data frame Grass gives the yield (10 * log10 dry-weight (g)) of eight grass Species in five replicates (Block) grown in sand culture at five levels of nitrogen.

Usage

data(Grass)

Format

A data frame with 40 observations on the following 7 variables.

Species

a factor with levels B.media D.glomerata F.ovina F.rubra H.pubesens K.cristata L.perenne P.bertolonii

Block

a factor with levels 1 2 3 4 5

N1

species yield at 1 ppm Nitrogen

N9

species yield at 9 ppm Nitrogen

N27

species yield at 27 ppm Nitrogen

N81

species yield at 81 ppm Nitrogen

N243

species yield at 243 ppm Nitrogen

Details

Nitrogen (NaNO3) levels were chosen to vary from what was expected to be from critically low to almost toxic. The amount of Nitrogen can be considered on a log3 scale, with levels 0, 2, 3, 4, 5. Gittins (1985, Ch. 11) treats these as equally spaced for the purpose of testing polynomial trends in Nitrogen level.

The data are also not truly multivariate, but rather a split-plot experimental design. For the purpose of exposition, he regards Species as the experimental unit, so that correlations among the responses refer to a composite representative of a species rather than to an individual exemplar.

Source

Gittins, R. (1985), Canonical Analysis: A Review with Applications in Ecology, Berlin: Springer-Verlag, Table A-5.

Examples


str(Grass)
grass.mod <- lm(cbind(N1,N9,N27,N81,N243) ~ Block + Species, data=Grass)


[Package candisc version 0.5-10 Index]