candisc-package(candisc)	R Documentation

Generalized Canonical Discriminant Analysis

Description

This package includes functions for computing and visualizing generalized canonical discriminant analyses for a multivariate linear model. They are designed to provide low-rank visualizations of terms in a mlm via the plot method and the heplots package.

Details

Package: 	 candisc

Type: 	 Package

Version: 	 0.5-10

Date: 	 2008-11-5

License: 	 GPL (>= 2)

The organization of functions in this package and the heplots package may change in a later version.

Author(s)

Michael Friendly and John Fox

Maintainer: John fox <jfox@mcmaster.ca>

References

Gittins, R. (1985). Canonical Analysis: A Review with Applications in Ecology, Berlin: Springer.

See Also

heplot

[Package candisc version 0.5-10 Index]