heplot.candiscList(candisc)	R Documentation

Canonical Discriminant HE plots

Description

These functions plot ellipses (or ellipsoids in 3D) in canonical discriminant space representing the hypothesis and error sums-of-squares-and-products matrices for terms in a multivariate linear model. They provide a low-rank 2D (or 3D) view of the effects for that term in the space of maximum discrimination.

Usage


## S3 method for class 'candiscList':
heplot(mod, term, ask = interactive(), graphics = TRUE, ...)

## S3 method for class 'candiscList':
heplot3d(mod, term, ask = interactive(), graphics = TRUE, ...)

Arguments

mod 	 A candiscList object for terms in a mlm

term 	 The name of one term to be plotted for the heplot and heplot3d methods. If not specified, one plot is produced for each term in the mlm object.

ask 	 If TRUE (the default), a menu of terms is presented; if ask is FALSE, canonical HE plots for all terms are produced.

graphics

... 	 Arguments to be passed down

Value

No useful value; used for the side-effect of producing canonical HE plots.

Author(s)

Michael Friendly and John Fox

References

Friendly, M. (2006). Data Ellipses, HE Plots and Reduced-Rank Displays for Multivariate Linear Models: SAS Software and Examples Journal of Statistical Software, 17(6), 1-42. http://www.jstatsoft.org/v17/i06/

Friendly, M. (2007). HE plots for Multivariate General Linear Models. Journal of Computational and Graphical Statistics, 16(2) 421-444. http://www.math.yorku.ca/SCS/Papers/jcgs-heplots.pdf

See Also

candisc, candiscList, heplot, heplot3d

[Package candisc version 0.5-10 Index]