Reflect Columns in an Object, reversing the sign of all elements

candisc objects have coefficients for the X and Y and weighted scores for these whose signs are arbitrary, in the sense that a given column can be reflected (multiplied by -1) without changing the fit. This function allows one to reflect any columns of the variable coefficients (and corresponding observation scores). This is often useful for interpreting a biplot, for example when a component (often the first) has all negative signs.

Usage

reflect(object, columns = 1:2, ...)

Arguments

object

An object whose columns are to be reflected

columns

a vector of indices of the columns to reflect

...

Unused

Value

The object with specified columns of the variable coefficients and observation scores multiplied by -1.

Author

Michael Friendly